requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.1
fastapi-limiter==0.1.5
ultralytics==8.2.30
opencv-python==4.9.0.80
Pillow==10.3.0
google-generativeai==0.6.0
python-dotenv==1.0.1
